Put formula rec into a ros pkg
