+++ intel/formula-recognition-medium-scan-0001/formula-recognition-medium-scan-0001-im2latex-encoder/vocab.json
@@ -1,0 +1,208 @@
+{
+    "id2sign": {
+        "0": "<s>",
+        "1": "<pad>",
+        "2": "</s>",
+        "3": "<unk>",
+        "4": "\\sqrt",
+        "5": "{",
+        "6": "\\cot",
+        "7": "9",
+        "8": "3",
+        "9": "+",
+        "10": "(",
+        "11": "7",
+        "12": ".",
+        "13": "1",
+        "14": "4",
+        "15": "T",
+        "16": "^",
+        "17": "5",
+        "18": "}",
+        "19": "Q",
+        "20": "8",
+        "21": ")",
+        "22": "X",
+        "23": "d",
+        "24": "-",
+        "25": "w",
+        "26": "6",
+        "27": "\\times",
+        "28": "H",
+        "29": "c",
+        "30": "x",
+        "31": "\\cos",
+        "32": "q",
+        "33": "z",
+        "34": "2",
+        "35": "\\theta",
+        "36": "F",
+        "37": "S",
+        "38": "j",
+        "39": "N",
+        "40": "l",
+        "41": "[",
+        "42": "\\arcsin",
+        "43": "u",
+        "44": "V",
+        "45": "U",
+        "46": "s",
+        "47": "]",
+        "48": "\\alpha",
+        "49": "f",
+        "50": "b",
+        "51": "m",
+        "52": "\\sin",
+        "53": "I",
+        "54": "g",
+        "55": "0",
+        "56": "n",
+        "57": "B",
+        "58": "i",
+        "59": "o",
+        "60": "\\arctan",
+        "61": "\\phi",
+        "62": "v",
+        "63": "\\sinh",
+        "64": "\\arccos",
+        "65": "G",
+        "66": "k",
+        "67": "Y",
+        "68": "h",
+        "69": "|",
+        "70": "t",
+        "71": "e",
+        "72": "W",
+        "73": "L",
+        "74": "y",
+        "75": "\\tanh",
+        "76": "\\beta",
+        "77": "K",
+        "78": "\\coth",
+        "79": "P",
+        "80": "Z",
+        "81": "a",
+        "82": "D",
+        "83": "=",
+        "84": "\\frac",
+        "85": "A",
+        "86": "O",
+        "87": "\\epsilon",
+        "88": "R",
+        "89": "C",
+        "90": "\\log",
+        "91": "p",
+        "92": "E",
+        "93": "\\pi",
+        "94": "M",
+        "95": "\\gamma",
+        "96": "\\tan",
+        "97": "J",
+        "98": "\\exp",
+        "99": "r",
+        "100": "\\cosh"
+    },
+    "sign2id": {
+        "(": 10,
+        ")": 21,
+        "+": 9,
+        "-": 24,
+        ".": 12,
+        "0": 55,
+        "1": 13,
+        "2": 34,
+        "3": 8,
+        "4": 14,
+        "5": 17,
+        "6": 26,
+        "7": 11,
+        "8": 20,
+        "9": 7,
+        "</s>": 2,
+        "<pad>": 1,
+        "<s>": 0,
+        "<unk>": 3,
+        "=": 83,
+        "A": 85,
+        "B": 57,
+        "C": 89,
+        "D": 82,
+        "E": 92,
+        "F": 36,
+        "G": 65,
+        "H": 28,
+        "I": 53,
+        "J": 97,
+        "K": 77,
+        "L": 73,
+        "M": 94,
+        "N": 39,
+        "O": 86,
+        "P": 79,
+        "Q": 19,
+        "R": 88,
+        "S": 37,
+        "T": 15,
+        "U": 45,
+        "V": 44,
+        "W": 72,
+        "X": 22,
+        "Y": 67,
+        "Z": 80,
+        "[": 41,
+        "\\alpha": 48,
+        "\\arccos": 64,
+        "\\arcsin": 42,
+        "\\arctan": 60,
+        "\\beta": 76,
+        "\\cos": 31,
+        "\\cosh": 100,
+        "\\cot": 6,
+        "\\coth": 78,
+        "\\epsilon": 87,
+        "\\exp": 98,
+        "\\frac": 84,
+        "\\gamma": 95,
+        "\\log": 90,
+        "\\phi": 61,
+        "\\pi": 93,
+        "\\sin": 52,
+        "\\sinh": 63,
+        "\\sqrt": 4,
+        "\\tan": 96,
+        "\\tanh": 75,
+        "\\theta": 35,
+        "\\times": 27,
+        "]": 47,
+        "^": 16,
+        "a": 81,
+        "b": 50,
+        "c": 29,
+        "d": 23,
+        "e": 71,
+        "f": 49,
+        "g": 54,
+        "h": 68,
+        "i": 58,
+        "j": 38,
+        "k": 66,
+        "l": 40,
+        "m": 51,
+        "n": 56,
+        "o": 59,
+        "p": 91,
+        "q": 32,
+        "r": 99,
+        "s": 46,
+        "t": 70,
+        "u": 43,
+        "v": 62,
+        "w": 25,
+        "x": 30,
+        "y": 74,
+        "z": 33,
+        "{": 5,
+        "|": 69,
+        "}": 18
+    }
+}
