+++ intel/formula-recognition-polynomials-handwritten-0001/formula-recognition-polynomials-handwritten-0001-encoder/vocab.json
@@ -1,0 +1,100 @@
+{
+    "id2sign": {
+        "0": "<s>",
+        "1": "<pad>",
+        "2": "</s>",
+        "3": "<unk>",
+        "4": "-",
+        "5": "4",
+        "6": "r",
+        "7": ".",
+        "8": "6",
+        "9": "c",
+        "10": "^",
+        "11": "{",
+        "12": "2",
+        "13": "}",
+        "14": "d",
+        "15": "0",
+        "16": "1",
+        "17": "5",
+        "18": "g",
+        "19": "9",
+        "20": "u",
+        "21": "7",
+        "22": "q",
+        "23": "+",
+        "24": "e",
+        "25": "v",
+        "26": "m",
+        "27": "=",
+        "28": "x",
+        "29": "a",
+        "30": "3",
+        "31": "i",
+        "32": "f",
+        "33": "w",
+        "34": "p",
+        "35": "b",
+        "36": "8",
+        "37": "s",
+        "38": "y",
+        "39": "l",
+        "40": "j",
+        "41": "k",
+        "42": "n",
+        "43": "z",
+        "44": "o",
+        "45": "t",
+        "46": "h"
+    },
+    "sign2id": {
+        "+": 23,
+        "-": 4,
+        ".": 7,
+        "0": 15,
+        "1": 16,
+        "2": 12,
+        "3": 30,
+        "4": 5,
+        "5": 17,
+        "6": 8,
+        "7": 21,
+        "8": 36,
+        "9": 19,
+        "</s>": 2,
+        "<pad>": 1,
+        "<s>": 0,
+        "<unk>": 3,
+        "=": 27,
+        "^": 10,
+        "a": 29,
+        "b": 35,
+        "c": 9,
+        "d": 14,
+        "e": 24,
+        "f": 32,
+        "g": 18,
+        "h": 46,
+        "i": 31,
+        "j": 40,
+        "k": 41,
+        "l": 39,
+        "m": 26,
+        "n": 42,
+        "o": 44,
+        "p": 34,
+        "q": 22,
+        "r": 6,
+        "s": 37,
+        "t": 45,
+        "u": 20,
+        "v": 25,
+        "w": 33,
+        "x": 28,
+        "y": 38,
+        "z": 43,
+        "{": 11,
+        "}": 13
+    }
+}
